review: dedup
